Break up lexer and parser into own modules
