+++ requirements.txt
@@ -1,1 +1,2 @@
 ply==3.8
+nose
